adds test for smoothie programming gpio
